moved StickerBuilder
