+++ examples/upload-story.example.ts
@@ -3,6 +3,8 @@ import Bluebird = require('bluebird');
 import { readFile } from 'fs';
 import { DateTime, Duration } from 'luxon';
 
+import { StickerBuilder} from '../src/sticker-builder';
+
 const ig = new IgApiClient();
 
 async function login() {
@@ -29,48 +31,48 @@ async function login() {
    *  bottom() aligns the sticker to the bottom
    *
    *  All of these are chainable e.g.:
-   *  ig.sticker.hashtag({ tagName: 'tag' }).scale(0.5).rotateDeg(90).center().left()
+   *  StickerBuilder.hashtag({ tagName: 'tag' }).scale(0.5).rotateDeg(90).center().left()
    */
 
   // these stickers are 'invisible' and not 're-rendered' in the app
   await ig.publish.story({
     file,
     // this creates a new config
-    stickerConfig: ig.sticker.init()
+    stickerConfig: new StickerBuilder()
     // these are all supported stickers
-      .add(ig.sticker.hashtag({
+      .add(StickerBuilder.hashtag({
         tagName: 'insta',
       }).center())
-      .add(ig.sticker.mention({
+      .add(StickerBuilder.mention({
         userId: ig.state.cookieUserId,
       }).center())
-      .add(ig.sticker.question({
+      .add(StickerBuilder.question({
         question: 'My Question',
       }).scale(0.5))
-      .add(ig.sticker.question({
+      .add(StickerBuilder.question({
         question: 'Music?',
         questionType: 'music',
       }))
-      .add(ig.sticker.countdown({
+      .add(StickerBuilder.countdown({
         text: 'My Countdown',
         endTs: DateTime.local().plus(Duration.fromObject({ hours: 1 })), // countdown finishes in 1h
       }))
-      .add(ig.sticker.chat({
+      .add(StickerBuilder.chat({
         text: 'Chat name',
       }))
-      .add(ig.sticker.location({
+      .add(StickerBuilder.location({
         locationId: (await ig.locationSearch.index(13, 37)).venues[0].external_id,
       }))
-      .add(ig.sticker.poll({
+      .add(StickerBuilder.poll({
         question: 'Question',
         tallies: [{ text: 'Left' }, { text: 'Right' }],
       }))
-      .add(ig.sticker.quiz({
+      .add(StickerBuilder.quiz({
         question: 'Question',
         options: ['0', '1', '2', '3'],
         correctAnswer: 1,
       }))
-      .add(ig.sticker.slider({
+      .add(StickerBuilder.slider({
         question: 'Question',
         emoji: '❤',
       }))
