+++ src/sticker-builder/index.ts
@@ -1,0 +1,2 @@
+export * from './sticker-builder';
+export * from './stickers';
